[Refactor][FE] 호텔 목록 조회 API 알맞은 클래스로 이동
