+++ frontend/src/lib/api/BusinessHotelApi.ts
@@ -8,6 +8,8 @@ import { PutHotelResponse } from "@/lib/types/hotel/PutHotelResponse";
 import { PostHotelRequest } from "@/lib/types/hotel/PostHotelRequest";
 import { RsData } from "./../types/RsData";
 import { useParams } from "next/navigation";
+import { PageDto } from "../types/PageDto";
+import { GetHotelResponse } from "../types/hotel/GetHotelResponse";
 
 const BASE_URL = "http://localhost:8080/api/hotels";
 
@@ -178,3 +180,45 @@ export const findAllHotelOptions =
       throw error;
     }
   };
+
+// 호텔 목록 획득 API
+export const getHotelList = async (
+  page: number,
+  pageSize: number,
+  filterName: string,
+  streetAddress: string,
+  checkInDate: string,
+  checkoutDate: string,
+  personal: number,
+  filterDirection?: string
+): Promise<PageDto<GetHotelResponse>> => {
+  const params = new URLSearchParams({
+    page: page.toString(),
+    pageSize: pageSize.toString(),
+    filterName: filterName,
+    streetAddress: streetAddress,
+    checkInDate: checkInDate,
+    checkoutDate: checkoutDate,
+    personal: personal.toString(),
+  });
+
+  if (filterDirection) {
+    params.append("filterDirection", filterDirection);
+  }
+
+  const url = `http://localhost:8080/api/hotels?${params.toString()}`;
+
+  try {
+    console.log(url);
+    const response = await fetch(url);
+    const rsData: RsData<PageDto<GetHotelResponse>> = await response.json();
+    console.log(rsData);
+    if (rsData.resultCode !== "200-1") {
+      throw new Error(rsData.msg);
+    }
+    return rsData.data;
+  } catch (error) {
+    console.error("Error:", error);
+    throw error;
+  }
+};
